requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
pydantic-ai[logfire]>=0.0.19
httpx>=0.28.0
python-dotenv>=1.0.1
logfire>=3.0.0
openai>=1.0.0
